Changed the way Outlook version is determined as there was a race condition. Outlook was not set up so I couldn't launch an email to determine the version of Outlook. I now find it out by the HKLM Path for Outlook and parse it.
I also added the automatic entry of Full Name and Initiials so Outlook will not prompt you.

There is also entries to set the "User Must Login For Mail" bit, but this is not yet complete. I am doing this for people that set this as it will prompt them to login if I launch Outlook. No biggie, but its annoying to me.